Observe changes to hasDismissedBlazeSectionOnMyStore to update Blaze section
