+++ WooCommerce/Classes/ViewRelated/Dashboard/Blaze/UserDefaults+Blaze.swift
@@ -1,0 +1,59 @@
+import Foundation
+import Combine
+
+extension UserDefaults {
+
+    /// Publisher to observe the changes to the dismissal state of the Blaze section on My Store screen.
+    var hasDismissedBlazeSectionOnMyStorePublisher: AnyPublisher<Bool, Never> {
+        ServiceLocator.stores.site
+            .combineLatest(self.publisher(for: \.hasDismissedBlazeSectionOnMyStore))
+            .map { [weak self] (site, hasDismissed) -> Bool in
+                guard let self, let site else {
+                    return false
+                }
+                return self.hasDismissedBlazeSectionOnMyStore(for: site.siteID)
+            }
+            .eraseToAnyPublisher()
+    }
+
+    /// Expose value for `hasDismissedBlazeSectionOnMyStore` to be observable through KVO.
+    @objc private var hasDismissedBlazeSectionOnMyStore: Bool {
+        get {
+            bool(forKey: Key.hasDismissedBlazeSectionOnMyStore.rawValue)
+        }
+        set {
+            set(newValue, forKey: Key.hasDismissedBlazeSectionOnMyStore.rawValue)
+        }
+    }
+
+    /// Checks if the Blaze section on My Store has been dismissed for a site.
+    ///
+    func hasDismissedBlazeSectionOnMyStore(for siteID: Int64) -> Bool {
+        let hasDismissed = self[.hasDismissedBlazeSectionOnMyStore] as? [String: Bool]
+        let idAsString = "\(siteID)"
+        return hasDismissed?[idAsString] == true
+    }
+
+    /// Marks the Blaze section on My Store as **not** dismissed for a site.
+    ///
+    func restoreBlazeSectionOnMyStore(for siteID: Int64) {
+        let idAsString = "\(siteID)"
+        guard var hasDismissed = self[.hasDismissedBlazeSectionOnMyStore] as? [String: Bool] else {
+            return
+        }
+        hasDismissed[idAsString] = false
+        self[.hasDismissedBlazeSectionOnMyStore] = hasDismissed
+    }
+
+    /// Marks the Blaze section on My Store as dismissed for a site.
+    ///
+    func setDismissedBlazeSectionOnMyStore(for siteID: Int64) {
+        let idAsString = "\(siteID)"
+        if var hasDismissed = self[.hasDismissedBlazeSectionOnMyStore] as? [String: Bool] {
+            hasDismissed[idAsString] = true
+            self[.hasDismissedBlazeSectionOnMyStore] = hasDismissed
+        } else {
+            self[.hasDismissedBlazeSectionOnMyStore] = [idAsString: true]
+        }
+    }
+}
